requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
pyyaml==6.0.2
python-dotenv==1.0.1
apscheduler==3.10.4
